requirements.txt
sentence-transformers==4.1.0
langchain_huggingface==0.1.2
streamlit==1.44.1
langchain_community==0.3.21
ollama==0.4.8
pypdf==5.4.0
faiss-cpu==1.10.0